+++ examples/lists-basic/main.tf
@@ -41,7 +41,7 @@ data "autocloud_blueprint_config" "ecs_custom_form" {
     // "ecs_cluster",                       // object({ arn = string name = string })
     //"ecs_subnet_ids",                  // list(string)
     "ecs_vpc_id",                        // string
-    "environment",                       // string
+    // "environment",                       // string
     "kms_key_id",                        // string
     "name",                              // string
     "additional_security_group_ids",     // list(string)
@@ -92,6 +92,25 @@ data "autocloud_blueprint_config" "ecs_custom_form" {
         "sales"
       ]
     })
+    
+    validation_rule {
+      rule          = "minLength"
+      value = 1
+      error_message = "min length is set to 1 element"
+    }
+    validation_rule {
+      rule          = "maxLength"
+      value = 5
+      error_message = "max length is set to 5 element"
+    }
+  }
+
+  variable {
+    name         = "environment"
+    display_name = "environment"
+    helper_text  = "select the environment"
+    type         = "raw"
+    value        = "var.ami"
   }
 
   variable {
@@ -143,6 +162,16 @@ data "autocloud_blueprint_config" "ecs_custom_form" {
       "3002",
       "3001",
     ])
+    validation_rule {
+      rule          = "minLength"
+      value = 1
+      error_message = "min length is set to 1 element"
+    }
+    validation_rule {
+      rule          = "maxLength"
+      value = 5
+      error_message = "max length is set to 5 element"
+    }
   }
 }