+++ internal/iac_catalog/blueprint/autocloud_blueprint_resource_test.go
@@ -122,11 +122,56 @@ func TestAccAutocloudBlueprint(t *testing.T) {
 }
 
 func TestAutocloudBlueprintHasAtMostOneGitConfigError(t *testing.T) {
-	t.SkipNow()
 	expectedError := `No more than 1 "git_config" blocks are allowed`
 	terraform := `resource "autocloud_blueprint" "bar" {
 		git_config {}
 		git_config {}
+	  }`
+	acctest.ValidateErrors(t, fmt.Errorf(expectedError), terraform)
+}
+
+func TestAutocloudBlueprintHasAtMostOneFileBlockError(t *testing.T) {
+	expectedError := `Insufficient file blocks`
+	terraform := `resource "autocloud_blueprint" "bar" {
+		git_config {}
+	  }`
+	acctest.ValidateErrors(t, fmt.Errorf(expectedError), terraform)
+}
+
+func TestAutocloudBlueprinntHasMissingAttributesOnFileBlockError(t *testing.T) {
+	expectedError := `file block should contain content or modules attributes`
+	terraform := `resource "autocloud_blueprint" "bar" {
+		file {
+			action      = "CREATE"
+			destination = "s3.tf"
+			variables = {}
+		}
+	  }`
+	acctest.ValidateErrors(t, fmt.Errorf(expectedError), terraform)
+}
+
+func TestAutocloudBlueprinntHasMissingModulesWithHeader(t *testing.T) {
+	expectedError := `modules can not be empty when using header or footer attributes`
+	terraform := `resource "autocloud_blueprint" "bar" {
+		file {
+			action      = "CREATE"
+			destination = "s3.tf"
+			variables = {}
+			header = "# header content"
+		}
+	  }`
+	acctest.ValidateErrors(t, fmt.Errorf(expectedError), terraform)
+}
+
+func TestAutocloudBlueprinntHasMissingModulesWitFooter(t *testing.T) {
+	expectedError := `modules can not be empty when using header or footer attributes`
+	terraform := `resource "autocloud_blueprint" "bar" {
+		file {
+			action      = "CREATE"
+			destination = "s3.tf"
+			variables = {}
+			footer = "# footer content"
+		}
 	  }`
 	acctest.ValidateErrors(t, fmt.Errorf(expectedError), terraform)
 }